requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
boto3==1.36.26
python-dotenv==1.0.1
requests==2.32.3
djangorestframework-simplejwt==5.4.0
google-auth==2.49.0
